[createAmpRender] refactor render functionality into createAmpRender     I'd like to be able to render an AMP page without getting hypernova involved. This is useful for validation and for integration with storybook.
